refactor: move utils to subdir
