+++ pixels-turbo/pixels-invoker-vhive/src/main/java/io/pixelsdb/pixels/invoker/vhive/utils/Utils.java
@@ -1,4 +1,4 @@
-package io.pixelsdb.pixels.invoker.vhive;
+package io.pixelsdb.pixels.invoker.vhive.utils;
 
 import com.alibaba.fastjson.JSON;
 import io.pixelsdb.pixels.executor.aggregation.FunctionType;
