+++ pixels-turbo/pixels-invoker-vhive/src/test/java/TestBroadcastChainJoinVhiveInvoker.java
@@ -2,7 +2,7 @@ import com.alibaba.fastjson.JSON;
 import io.pixelsdb.pixels.common.physical.Storage;
 import io.pixelsdb.pixels.common.turbo.InvokerFactory;
 import io.pixelsdb.pixels.common.turbo.WorkerType;
-import io.pixelsdb.pixels.invoker.vhive.Utils;
+import io.pixelsdb.pixels.invoker.vhive.utils.Utils;
 import io.pixelsdb.pixels.planner.plan.physical.domain.StorageInfo;
 import io.pixelsdb.pixels.planner.plan.physical.output.JoinOutput;
 import org.junit.Test;
